requirements.txt
streamlit
openai-whisper==20230918
pyannote.audio

